# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for fetching country information.",
        "title": "Country API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/api/countries/search": {
            "get": {
                "description": "Fetches details of a country using its name.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Countries"
                ],
                "summary": "Search country by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Country Name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved country details",
                        "schema": {
                            "$ref": "#/definitions/model.Country"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Missing country name",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Country": {
            "type": "object",
            "properties": {
                "capital": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "currencies": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/model.Currency"
                    }
                },
                "name": {
                    "$ref": "#/definitions/model.Name"
                },
                "population": {
                    "type": "integer"
                }
            }
        },
        "model.Currency": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "model.Name": {
            "type": "object",
            "properties": {
                "common": {
                    "type": "string"
                }
            }
        }
    }
}